requirements.txt
azure-identity
graphrag

openai

faiss-cpu
numpy

fastmcp
pypdf